+++ deltatech_website_product_code/tests/__init__.py
@@ -1,0 +1,5 @@
+# ©  2026 Deltatech
+#              Dorin Hongu <dhongu(@)gmail(.)com
+# See README.rst file on addons root folder for license details
+
+from . import test_exact_phrase_search